Phase 3 Wave 7: Move 27 files to commonMain + KClass conversion + JVM dep removal
* Strip @JvmStatic/@JvmField from 8 high-value blocker files

Unblocks ~116 additional files for commonMain migration:
- TreeReference: rootRef(), selfRef() — unblocks 77 files
- XPathParseTool: parseXPath() — unblocks 21 files
- FunctionUtils: 18 companion methods — unblocks 18 files
- XPathFuncExpr: name, args, expectedArgCount fields
- Parser: parse() method
- Base64, JsonUtils, InstanceUtils: companion methods

Updated 21 Java callers to use Companion/INSTANCE accessor syntax.


* Move 18 leaf files from src/main/java to src/commonMain/kotlin

Move files with zero JVM dependencies and no transitive deps on
blocked files to commonMain. These are "leaf" files in the dependency
graph - data classes, interfaces, and simple models that only depend
on things already in commonMain.

Files moved:
- cases: Ledger, CaseIndex, DualTableSingleMatchModelQuerySet
- network: MemoryBitCache
- engine/cases: RecordObjectCache, RecordSetResultCache
- suite/model: AndroidPackageDependency, ComputedDatum, Credential,
  EndpointAction, EndpointArgument, FormIdDatum, PropertySetter,
  SessionDatum
- javarosa/model: QuestionString, UploadQuestionExtension,
  InvalidDataException
- services/locale: TableLocaleSource

commonMain: 157 → 175 files
src/main/java: 488 → 470 .kt files


* Remove JVM System.out/err and java.util.Objects dependencies from 17 files

Replace System.out.println → println() (Kotlin stdlib, cross-platform)
Replace System.err.println → platformStdErrPrintln() (expect/actual)
Replace java.util.Objects.requireNonNull → requireNotNull() (Kotlin stdlib)
Replace javaClass.name → ::class.simpleName (KMP-compatible)

These were trivial JVM dependencies preventing future commonMain migration.
29 replacements across 17 files.


* Replace Class<T> with KClass<T> in QueryContext/QueryCacheHost

Convert the query cache API from JVM-only Class<T> to KMP-compatible
KClass<T> + factory lambda. This removes ::class.java from 10 files
and unblocks QueryContext for commonMain migration.

- QueryCacheHost: HashMap<Class<*>, QueryCache> → HashMap<KClass<*>, QueryCache>
- getQueryCache(Class<T>) → getQueryCache(KClass<T>, () -> T)
- Callers: SomeClass::class.java → SomeClass::class + { SomeClass() }
- Java-compatible overload kept for test interop
- QueryCacheTests.java → QueryCacheTests.kt


* Replace JVM dependencies: add kotlin.jvm imports, Math→kotlin.math, Double.valueOf removal

- Add explicit kotlin.jvm.JvmField/JvmStatic/JvmOverloads imports to 137 files
  (required for commonMain compilation where these aren't auto-imported)
- Replace java.lang.Math with kotlin.math equivalents in 22 files
- Replace java.lang.Double.valueOf() with Kotlin idioms in 12 files
- All 710+ JVM tests pass


* Move Property.kt to commonMain (after System.out.println removal)


* Convert 8 Java files to Kotlin: exceptions, interfaces, reference factories

- UnfullfilledRequirementsException (66 refs as unresolved blocker)
- InvalidReferenceException (26 refs)
- TransactionParserFactory → fun interface for SAM conversion (19 refs)
- ActionableInvalidStructureException, InvalidCasePropertyLengthException
- IModule, ReferenceFactory, PrefixedRootFactory, ResourceReferenceFactory
- All 710+ JVM tests pass


* Move 5 more files to commonMain (181 total)

QueryCacheHost, IModule, InvalidReferenceException,
InvalidCasePropertyLengthException, UnfullfilledRequirementsException


* Convert storage interfaces from Class<*> to KClass<*> for KMP

- IStorageUtilityIndexed.getPrototype(): Class<*> → KClass<*>
- IStorageIndexedFactory.newStorage(name, type): Class<*> → KClass<*>
- StorageManager.registerStorage(key, type): KClass<*> primary, Class<*> overload
- DummyIndexedStorageUtility: KClass + factory lambda, Class<T> compat constructor
- CommCarePlatform: Property::class.java → Property::class
- CLI Java callers updated with JvmClassMappingKt.getKotlinClass()
- All 710+ JVM tests pass


* Move 3 more files to commonMain (184 total)

IStorageUtilityIndexed, IStorageIndexedFactory, PropertyManager
- Storage interfaces now KMP-compatible after KClass conversion


* Move QueryContext Class<T> overload to jvmMain extension

Extract the Java-compatible getQueryCache(Class<T>) from QueryContext
to a JVM extension function in jvmMain, making QueryContext itself
free of JVM dependencies and eligible for commonMain.


---------

+++ commcare-core/src/main/java/org/javarosa/xpath/expr/XPathArithExpr.kt
@@ -22,7 +22,7 @@ class XPathArithExpr : XPathBinaryOpExpr {
             DIVIDE -> result = aval / bval
             MODULO -> result = aval % bval
         }
-        return java.lang.Double.valueOf(result)
+        return result
     }
 
     override fun toString(): String {
